[REFACTOR] 패키지 이전: Avatar 관련 Repository -> avatar/model
