+++ src/main/java/com/wakeUpTogetUp/togetUp/api/users/UserAvatarService.java
@@ -2,9 +2,10 @@ package com.wakeUpTogetUp.togetUp.api.users;
 
 import static com.wakeUpTogetUp.togetUp.common.Constant.DEFAULT_AVATAR_ID;
 
-import com.wakeUpTogetUp.togetUp.api.avatar.AvatarRepository;
-import com.wakeUpTogetUp.togetUp.api.avatar.model.Avatar;
 import com.wakeUpTogetUp.togetUp.api.avatar.dto.response.UserAvatarResponse;
+import com.wakeUpTogetUp.togetUp.api.avatar.model.Avatar;
+import com.wakeUpTogetUp.togetUp.api.avatar.repository.AvatarRepository;
+import com.wakeUpTogetUp.togetUp.api.avatar.repository.UserAvatarRepository;
 import com.wakeUpTogetUp.togetUp.api.users.model.User;
 import com.wakeUpTogetUp.togetUp.api.users.model.UserAvatar;
 import com.wakeUpTogetUp.togetUp.common.Status;
